Test de creacion de materiales
